requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
flask-login>=0.6.3
pandas>=2.0.0
pyarrow>=14.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
chardet>=5.0.0
numpy>=1.24.0
structlog>=23.0.0
plotly>=5.0.0
flask-caching>=2.0.0
moto[s3]>=5.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
